GraphMeta, node_match(), get_edges()
node_match() - a function enabling comparing two nodes for semantic
equality
get_edges() - a function enabling getting edges of self as well as edges
from neighbors.
GraphMeta - a class inner to BaseClass that determines which attributes
to examine for computing edges, and which attributes/properties to
compare for deciding node matches.